Update iblindsV3_10.groovy
Updated the logs to show device name to help identify which blinds are reporting status.  Removed tabs from file.  Removed comma from refresh() method that was adding an additional command with zero length.   Changed log.debug to log.info to be more consistent with other drivers behavior.

Biggest change fixing battery reporting.  Now battery level is reported in the log.  Also, getBattery() was changed to send the command a as a single multi-command.  Although the previous code looked correct, it was causing an error and not correctly being handled in the zwave battery report handler.  The error was displayed in the log file and appeared to be coming from the firmware 3.06 while running Hubitat C-7 Version 2.2.8.147